requirements.txt
numpy
scipy
scikit-image
matplotlib
pillow
